backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.9.0
django-filter==25.1
django-rest-passwordreset==1.5.0
channels==4.3.2
channels_redis==4.3.0
redis==5.2.0
python-dateutil==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
dj-database-url==2.3.0
gunicorn==23.0.0
whitenoise==6.8.2
requests==2.32.3
openpyxl==3.1.5
XlsxWriter==3.2.0
pandas==2.3.3
Pillow==12.1.0
mysqlclient==2.2.4
PyMySQL==1.1.2
psycopg2-binary==2.9.10
django-jazzmin==3.0.1
django-debug-toolbar==4.4.6